Updated check point behavior: player start now works respawning no longer requires restarting scene.
Made game instance and set it to default.
Moved everything to do with saving or that needs to be saved to the Game instance.

Added empty levels the crystal city and death factory
Made volume that moves the player to the next level and records completing the level the volume is in.

Updated main menu to have a level select
Added icons to show if a level has been completed and if the power of the level has been collected.